Rename search repo to elasticsearch package
